requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
timm==1.0.24
pillow==12.0.0
numpy==2.2.6
albumentations==2.0.8
tqdm==4.67.3